Renaming contains to pages.
