+++ client/containers/AccountsPage/index.js
@@ -1,50 +1,18 @@
-/* global fetch */
+
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 import Header from '../../components/Header';
 import Footer from '../../components/Footer';
 import style from './style.css';
 
-class AppList extends Component {
-  constructor (...args) {
-    super(...args);
-    this.state = {
-      apps: []
-    };
-  }
-
-  componentDidMount () {
-    // TODO [todr] move to state
-    fetch('/api/apps', {
-      credentials: 'same-origin'
-    })
-      .then((res) => res.json())
-      .then((data) => this.setState({
-        apps: data
-      }));
-  }
-
-  renderApps () {
-    if (!this.state.apps) {
-      return;
-    }
-    return this.state.apps.map((app) => (
-      <li key={app}>
-        <a href={`/${app}/`}>{app}</a>
-      </li>
-    ));
-  }
-
+class AccountsPage extends Component {
   render () {
     return (
       <div className={style.normal}>
         <Header nodeName={this.props.status.name} />
         <div className='dapp-flex-content'>
           <main className='dapp-content'>
-            <h1>List of installed apps</h1>
-            <ul>
-            {this.renderApps()}
-            </ul>
+            <h1>Accounts</h1>
           </main>
         </div>
         <Footer version={this.props.status.version} />
@@ -52,7 +20,7 @@ class AppList extends Component {
     );
   }
 }
-AppList.propTypes = {
+AccountsPage.propTypes = {
   status: PropTypes.object.isRequired
 };
 
@@ -67,4 +35,4 @@ function mapDispatchToProps (dispatch) {
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(AppList);
+)(AccountsPage);
