+++ client/containers/AppListPage/index.js
@@ -1,18 +1,50 @@
-
+/* global fetch */
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 import Header from '../../components/Header';
 import Footer from '../../components/Footer';
 import style from './style.css';
 
-class Accounts extends Component {
+class AppListPage extends Component {
+  constructor (...args) {
+    super(...args);
+    this.state = {
+      apps: []
+    };
+  }
+
+  componentDidMount () {
+    // TODO [todr] move to state
+    fetch('/api/apps', {
+      credentials: 'same-origin'
+    })
+      .then((res) => res.json())
+      .then((data) => this.setState({
+        apps: data
+      }));
+  }
+
+  renderApps () {
+    if (!this.state.apps) {
+      return;
+    }
+    return this.state.apps.map((app) => (
+      <li key={app}>
+        <a href={`/${app}/`}>{app}</a>
+      </li>
+    ));
+  }
+
   render () {
     return (
       <div className={style.normal}>
         <Header nodeName={this.props.status.name} />
         <div className='dapp-flex-content'>
           <main className='dapp-content'>
-            <h1>Accounts</h1>
+            <h1>List of installed apps</h1>
+            <ul>
+            {this.renderApps()}
+            </ul>
           </main>
         </div>
         <Footer version={this.props.status.version} />
@@ -20,7 +52,7 @@ class Accounts extends Component {
     );
   }
 }
-Accounts.propTypes = {
+AppListPage.propTypes = {
   status: PropTypes.object.isRequired
 };
 
@@ -35,4 +67,4 @@ function mapDispatchToProps (dispatch) {
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(Accounts);
+)(AppListPage);
